Fix MSVC compiler warnings (replace most M_PI with (float)M_PI)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7404 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/MapReader.cpp
@@ -1110,7 +1110,7 @@ int CXMLReader::ReadNonEntities(XMBElement parent, double end_time)
 			if (unit)
 			{
 				CMatrix3D m;
-				m.SetYRotation(Orientation + M_PI);
+				m.SetYRotation(Orientation + (float)M_PI);
 				m.Translate(Position);
 				unit->GetModel()->SetTransform(m);
